시작화면 생성 and 모듈화 성공
