+++ SwTeamProject/StartScreen.h
@@ -1,0 +1,8 @@
+#pragma once
+#include "Common.h"
+
+#define E 101
+#define H 104
+#define Q 113
+void printStartScreen(void);
+int getKeyInput(void);
