--- SwTeamProject/Track.h
@@ -1,7 +1,0 @@
-#pragma once
-#include "characters.h"
-#include "cursor.h"
-
-int getDistance(int, int);
-void setNpcPrintPos(NPC*);
-void updateNpcPos(player*, NPC*);
