--- SwTeamProject/cursor.c
@@ -1,27 +1,0 @@
-#include "cursor.h"
-
-void SetCurrentCursorPos(int x, int y)
-{
-	COORD pos = { x,y };
-	SetConsoleCursorPosition(GetStdHandle(STD_OUTPUT_HANDLE), pos);
-}
-
-COORD GetCurrentCursorPos(void)
-{
-	COORD curPoint;
-	CONSOLE_SCREEN_BUFFER_INFO curInfo;
-
-	GetConsoleScreenBufferInfo(GetStdHandle(STD_OUTPUT_HANDLE), &curInfo);
-	curPoint.X = curInfo.dwCursorPosition.X;
-	curPoint.Y = curInfo.dwCursorPosition.Y;
-
-	return curPoint;
-}
-
-void RemoveCursor(void)
-{
-	CONSOLE_CURSOR_INFO curInfo;
-	GetConsoleCursorInfo(GetStdHandle(STD_OUTPUT_HANDLE), &curInfo);
-	curInfo.bVisible = 0;
-	SetConsoleCursorInfo(GetStdHandle(STD_OUTPUT_HANDLE), &curInfo);
-}
