--- SwTeamProject/cursor.h
@@ -1,6 +1,0 @@
-#pragma once
-#include <Windows.h>
-
-void SetCurrentCursorPos(int x, int y);
-COORD GetCurrentCursorPos(void);
-void RemoveCursor(void);
